+++ src/test/java/com/vaadin/ui/dialog/tests/DialogIT.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.ui.dialog;
+package com.vaadin.ui.dialog.tests;
 
 import org.junit.Assert;
 import org.junit.Test;
@@ -23,6 +23,7 @@ import org.openqa.selenium.interactions.Actions;
 
 import com.vaadin.flow.demo.ComponentDemoTest;
 import com.vaadin.testbench.By;
+import com.vaadin.ui.dialog.demo.DialogView;
 
 /**
  * Integration tests for the {@link DialogView}.
